added tests for getMochaOpts
