+++ lib/cli.js
@@ -2,11 +2,12 @@
 "use strict";
 
 var path = require("path");
+var getMochaOpts = require("./getMochaOpts");
 var programmaticRunner = require("./programmaticRunner");
 
 var filePath = getAdapterFilePath();
 var adapter = adapterObjectFromFilePath(filePath);
-var mochaOpts = getMochaOpts();
+var mochaOpts = getMochaOpts(process.argv.slice(3));
 programmaticRunner(adapter, mochaOpts, function (err) {
     if (err) {
         process.exit(err.failures || -1);
@@ -19,24 +20,6 @@ function getAdapterFilePath() {
     } else {
         throw new Error("Specify your adapter file as an argument.");
     }
-}
-
-function getMochaOpts() {
-    var rawOpts = process.argv.slice(3);
-    var opts = {};
-
-    rawOpts.join(" ").split("--").forEach(function (opt) {
-        var optSplit = opt.split(" ");
-
-        var key = optSplit[0];
-        var value = optSplit[1] || true;
-
-        if (key) {
-            opts[key] = value;
-        }
-    });
-
-    return opts;
 }
 
 function adapterObjectFromFilePath(filePath) {
